04. 3 Goals Progress/01. Accounting/requirements.txt
google-genai>=1.0.0
pypdf>=5.1.0
